+++ src/astro/FunctionalPalace.ts
@@ -1,6 +1,6 @@
 import { Palace } from '../data/types';
-import { StarName } from '../i18n';
-import { hasOneOfStars, hasStars, notHaveStars } from './analyzer';
+import { Mutagen, StarName } from '../i18n';
+import { hasMutagenInPlace, hasOneOfStars, hasStars, notHaveMutagenInPalce, notHaveStars } from './analyzer';
 
 export interface IFunctionalPalace extends Palace {
   /**
@@ -32,6 +32,26 @@ export interface IFunctionalPalace extends Palace {
    * @returns true | false
    */
   hasOneOf: (stars: StarName[]) => boolean;
+
+  /**
+   * 判断宫位内是否有生年四化
+   *
+   * @version v1.2.0
+   *
+   * @param mutagen 四化名称【禄｜权｜科｜忌】
+   * @returns true | false
+   */
+  hasMutagen: (mutagen: Mutagen) => boolean;
+
+  /**
+   * 判断宫位内是否没有生年四化
+   *
+   * @version v1.2.0
+   *
+   * @param mutagen 四化名称【禄｜权｜科｜忌】
+   * @returns true | false
+   */
+  notHaveMutagen: (mutagen: Mutagen) => boolean;
 }
 
 export default class FunctionalPalace implements IFunctionalPalace {
@@ -72,4 +92,6 @@ export default class FunctionalPalace implements IFunctionalPalace {
   has = (stars: StarName[]): boolean => hasStars(this, stars);
   notHave = (stars: StarName[]): boolean => notHaveStars(this, stars);
   hasOneOf = (stars: StarName[]): boolean => hasOneOfStars(this, stars);
+  hasMutagen = (mutagen: Mutagen): boolean => hasMutagenInPlace(this, mutagen);
+  notHaveMutagen = (mutagen: Mutagen): boolean => notHaveMutagenInPalce(this, mutagen);
 }